# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/fransfilastap/kontak"
    },
    "title": "Kontak API",
    "version": "1.0",
    "description": "WhatsApp-to-REST-API bridge",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/admin/inbox/{client_id}/schedule": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Schedule a message to be sent at a specific time",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "Schedule message",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Message data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pkg_http.ScheduleMessageRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/inbox/{client_id}/threads": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get all message threads for a device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "List threads",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/inbox/{client_id}/threads/send": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Send a text message to a new recipient (creates thread if needed)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "Send new message",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Message data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pkg_http.SendNewMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/inbox/{client_id}/threads/{chat_jid}/media": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Send a media file (image, video, audio, document) to an existing thread",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "Send media message",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Chat JID",
            "name": "chat_jid",
            "in": "path",
            "required": true
          },
          {
            "type": "file",
            "description": "Media file",
            "name": "file",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/inbox/{client_id}/threads/{chat_jid}/messages": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get messages for a specific thread",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "List thread messages",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Chat JID",
            "name": "chat_jid",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/inbox/{client_id}/threads/{chat_jid}/read": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Mark all messages in a thread as read",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "Mark thread as read",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Chat JID",
            "name": "chat_jid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/pkg_http.GenericResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/inbox/{client_id}/threads/{chat_jid}/send": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Send a text message to an existing thread",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "inbox"
        ],
        "summary": "Send message to thread",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Chat JID",
            "name": "chat_jid",
            "in": "path",
            "required": true
          },
          {
            "description": "Message data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pkg_http.SendInboxMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "pkg_http.ErrorResponse": {
      "type": "object",
      "properties": {
        "context": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.GenericResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.ScheduleMessageRequest": {
      "type": "object",
      "required": [
        "scheduled_at",
        "text",
        "to"
      ],
      "properties": {
        "scheduled_at": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "to": {
          "type": "string"
        }
      }
    },
    "pkg_http.SendInboxMessageRequest": {
      "type": "object",
      "required": [
        "text"
      ],
      "properties": {
        "text": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.SendNewMessageRequest": {
      "type": "object",
      "required": [
        "text",
        "to"
      ],
      "properties": {
        "text": {
          "type": "string"
        },
        "to": {
          "type": "string"
        }
      }
    },
...
  },
  "host": "localhost:8080",
  "basePath": "/"
}